P7 §8 Done + sync migration 0007 to the applied formula (round -> ::numeric)
Consolidates the verified P7 state (WhatsApp/voice farmer channel) into
MASTER.md §8 with test evidence + notes for the next AI, and aligns the
migration file's nearest_village() distance rounding with the form actually
applied live (round(double,integer) is not a PostgreSQL function; cast the
km value to numeric first).

+++ supabase/migrations/0007_channel.sql
@@ -73,7 +73,7 @@ returns table (name text, taluka text, district text, distance_km double precisi
 language sql stable set search_path = public
 as $$
   select v.name, v.taluka, v.district,
-         round(st_distance(v.geo::geography, st_setsrid(st_makepoint(p_lng, p_lat), 4326)::geography) / 1000.0, 2) as distance_km
+         round((st_distance(v.geo::geography, st_setsrid(st_makepoint(p_lng, p_lat), 4326)::geography) / 1000.0)::numeric, 2) as distance_km
   from public.villages v
   order by v.geo <-> st_setsrid(st_makepoint(p_lng, p_lat), 4326)::geography
   limit 1;
